isochron: automatically determine the destination MAC address
If the sender has a stats socket towards the receiver, it can query the
MAC address that the receiver would like to use for this test. This
should make isochron scripts more portable for the basic case there the
destination MAC address is the unicast address of the station that the
receiver is listening on.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>