isochron: send: print the RX deadline delta in reverse if packets arrived early
Currently, if the isochron sender doesn't use SO_TXTIME or tc-taprio,
chances are that packets will arrive way earlier than their scheduled
time, because the MAC will not delay them.

This results in us printing a negative value for the hardware RX
deadline delta. This is confusing for the user.

Detect this case and recompute the statistics interpreted in reverse.
Also tell the user what the metric means, exactly.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>