isochron: print the sender latency
An interesting metric for a TSN endpoint is how fast it can wrap up
sending a packet compared to the length of the cycle. So let's print
exactly that, the percentage of time that the sender spends waking up,
and the percentage it spends running (this includes the kernel time,
driver, packet scheduler, etc etc). We consider that processing for a
packet stops when the driver passes it to the TX ring, which is not
entirely true (the TX completion consumes time too), but as far as
communication is concerned, time spent after the packet transmission is
not so relevant.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>