isochron: send: fix subtly invalid software TX timestamps
We calculate the time it takes to send a packet based on the delta
between the application wakeup time and the software TX timestamp, but
even with perfect sync and a ridiculously large cycle time, one can see
that isochron apparently thinks we've exceeded the alloted time,
although we didn't.

Sending one packet takes on average 90.229% of the cycle time (min 0.873% max 100.200%)
Waking up takes on average 0.152% of the cycle time (min 0.144% max 0.182%)

The Linux TX timestamping UAPI is garbage and my understanding of it
was even worse.

Documentation/networking/timestamping.rst says:

| Note that if the SO_TIMESTAMP or SO_TIMESTAMPNS option is enabled
| together with SO_TIMESTAMPING using SOF_TIMESTAMPING_SOFTWARE, a false
| software timestamp will be generated in the recvmsg() call and passed
| in ts[0] when a real software timestamp is missing. This happens also
| on hardware transmit timestamps.

which is basically what is happening here. We aren't seeing a software
timestamp filled in by the driver, but instead a fake one created by the
kernel to make us happy and shut up. Even though one can see that the
software TX timestamp is larger than the hardware TX timestamp, which is
absurd.

So if we want reliable SOF_TIMESTAMPING_OPT_TX_SWHW behavior (both
software and hardware timestamps for the same skb), we should drop
support for retrieving software timestamps through SO_TIMESTAMPNS and go
all the way with SO_TIMESTAMPING.

Except that SO_TIMESTAMPING has other crap of its own:

| SOF_TIMESTAMPING_OPT_TX_SWHW:
|   Request both hardware and software timestamps for outgoing packets
|   when SOF_TIMESTAMPING_TX_HARDWARE and SOF_TIMESTAMPING_TX_SOFTWARE
|   are enabled at the same time. If both timestamps are generated,
|   two separate messages will be looped to the socket's error queue,
|   each containing just one timestamp.

And we aren't yet prepared to deal with multiple sk_receive() calls for
the same packet, to retrieve multiple TX timestamps of it. Although this
is what this patch does.

Because the TX timestamps can arrive in any order (the TX timestamps of
the same packet, and also TX timestamps of different packets, because we
don't block during the hot path waiting for a TX timestamp), we also
employ this other feature:

| SOF_TIMESTAMPING_OPT_ID:
|   Generate a unique identifier along with each packet. A process can
|   have multiple concurrent timestamping requests outstanding. Packets
|   can be reordered in the transmit path, for instance in the packet
|   scheduler. In that case timestamps will be queued onto the error
|   queue out of order from the original send() calls. It is not always
|   possible to uniquely match timestamps to the original send() calls
|   based on timestamp order or payload inspection alone, then.
|
|   This option associates each packet at send() with a unique
|   identifier and returns that along with the timestamp. The identifier
|   is derived from a per-socket u32 counter (that wraps). For datagram
|   sockets, the counter increments with each sent packet. For stream
|   sockets, it increments with every byte.

So we need to change the way in which we log data in the sender.
Because the kernel will deliver us control messages containing only one
timestamp at a time (either software or hardware), we need to live-patch
log buffer entries instead of logging a packet once and being done with it.
We then count packets as fully timestamped when both the software and
the hardware TX timestamp has been provided, and stop waiting when all
packets have been fully timestamped.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>