isochron: send: don't attempt to print the packet log on errors
We might kill the sender before the test has even started (for example
while it's still waiting for sync to be established), and in that case,
the log printing is located in prog_teardown() which is quite bad
because there's nothing to print, but it will still calculate the
statistics on invalid data and print a bunch of "not a number" values.

Avoid that by moving the packet log printing into a function that is not
called on error teardown.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>