isochron: keep packets sorted by seqid in the log buffer
Currently, calculating statistics takes a very long time because for
each send packet must be looked up by {seqid, txtime} in the receive log
buffer, then we remove it by memmoving every other receive log buffer
entry to the left.

We can reduce the complexity if we always log a packet at an index in
the log that depends on the seqid. Since seqid 0 is reserved, we always
log at an index equal to seqid - 1, and finding an entry with seqid 0
means that there is no packet logged there.

Before this patch, calculating the statistics for --num-frames 50000:

Processing the stats took 80.985658347 seconds

After this patch, same test:

Processing the stats took 0.502869764 seconds

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>