isochron: send: check sync again at the end of the test
It may happen that the sender and receiver were in sync when the test
started, but they lost sync by the time it ended.

By design I don't want isochron to monitor the sync during the actual
test itself, that process is already very time sensitive as it is, and
the ptpmon/sysmon API is not asynchronous or subscription-based, so it
would basically break the test.

But let's at least check again at the end that synchronization quality
is still good. For example, in buggy systems, isochron itself might
break it, due to a high scheduling priority, quantity of timestamped
packets, or just network load.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>