isochron: send: calculate the driver latency using scheduler software timestamps
Now that we took the time and effort to convert isochron to 100%
SO_TIMESTAMPING API, we can just as well retrieve a second TX software
timestamp, which is taken in __dev_queue_xmit().

From this value we can derive the driver's latency, which is the
software timestamp collected with skb_tx_tstamp(), minus the scheduler
timestamp.

Note that this alters struct isochron_send_pkt_data, so it breaks
compatibility with current isochron log version. However I'm not bumping
the version just now, there are some more changes I want to make and I'd
like to bump the version just once.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>