isochron: move log related stuff to log.{c,h}
For better modularity, let's move away from dumping everything that
isn't strictly sender-related or receiver-related to common.c.

Let's start with the packet log buffer stuff, create a log.c file that
deals with that.

As usual, keep inside log.c the structure definitions that don't need to
be exported.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>