isochron: report: save and view logs from the filesystem
The isochron sender is becoming less and less focused, it is gathering
more data and trying to print it in a way which may or may not be
meaningful to the end user, like for example the metric statistics that
we extract and print are pretty much arbitrary.

Furthermore, right now we don't even print all the data we gather, for
example we omit printing the software timestamps per packet even though
we calculate statistics based on them. This is because printing
everything we have would fill even a very wide line of text, which is
unpleasant for a user to say the least.

So it looks like it may be just simpler for the sender to log what it
has collected to the filesystem, then leave the task for a different
program to interpret that data. That different program can receive input
from the user as to what data to filter for and how to print it. That is
the "isochron record" sub-program.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>