isochron: drop the "stats" sub-program
The "stats" program was created in haste due to the need to recalculate
statistics on a subset of the collected data. Since at the time, the
only entity who printed the data and calculated the statistics for the
predefined metrics was the sender, yet we had to look at the data before
knowing which parts of it needed to be cropped out, the result was that
post-processing of sender logs needed to be done.

The problem is that text file processing is not really great, especially
since the isochron sender print format is not stable and is not, in
fact, even printing all the data it gathers.

Additionally, since basically the "stats" program needed to be fed the
actual output of the "send" program, it also needed to be told what
runtime parameters (cycle time, advance time etc) were used for that
test, to calculate something useful out of them.

So the "stats" program is the wrong solution to the problem. Let's focus
all our attention on the "report" program, which requires a lot less
arguments (since the sender records the parameters of a test in the
actual output file), and can decide what to filter out depending on what
the user wants (to be added in future patches).

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>