isochron: argparser: avoid unterminated string based attacks
strncpy() is a dangerous function and does not ensure the string is NULL
terminated, which in turn means that if the user supplies a string
larger than our predefined maximum, it will be truncated but we might
still read data beyond the string length.

Avoid that by simply erroring out if truncation would occur.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>