isochron: report: add a printf format specifier
Isochron started acquiring more data than would be useful to print by
default, and there isn't a good way to keep growing that output (plus,
we can't guarantee that the output format remains stable).

Also, the metrics computed by isochron may not be the metrics needed by
the user.

So add an option for the user to choose what gets printed, and in what
format. Computing user-defined metrics is way beyond the scope of this
program, since doing so would imply having an arithmetic evaluator
built-in. But as long as we give an option to print an internal variable
using all the formats that make sense (aka print a time either in the
human-friendly "sec.nsec" format, or in the computer-friendly integer
format), then calculating the user-defined metrics in an external
program, like a spreadsheet editor, should not be a problem:

isochron report \
	--input-file isochron.log \
	--printf-format "%u,%d,%d" \
	--printf-args "qts" \
	> isochron.csv

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>