isochron: argparser: add a dedicated parser for unsigned longs
There are justified compiler warnings on 32 bit systems due to the fact
that we compare signed with unsigned variables, and this is because the
command line arguments received from the user are stored in signed long
integers.

Add an option to the argparser to store as unsigned long where that
makes sense.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>