isochron: log: reduce the number of calls to strlen
For a data set of 10000000 packets, the following command:

isochron report \
	--input-file isochron-long.log \
	--printf-format "%d,%d,%d,%d,%d,%d,%d,%d,%d,%d,%d" \
	--printf-args "ABCHWSwtTrR" \
	> /dev/null

takes:

- on x86_64: 11.872s before this patch, 11.864s afterwards
- on armv8: 2m56.279s before this patch, 2m51.292s afterwards

(ignoring gross oversimplifications such as the fact that the armv8
system has an UHS-1 SD card for storage and the x86_64 system has an
NVME SSD)

it is absolutely clear that the bottleneck for this command is actually
reading the entire log file to memory and handling the page faults.

But in any case, we can avoid calling strlen() twice for small strings,
even if it's insignificant in the grand scheme.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>