isochron: prepare code base for multi threading by not using strerror()
The strerror() libc function is marked by man pthreads(7) as being not
thread-safe, due to the fact that a static temporary buffer is used to
print the value of errno in case the error code is unknown. There exists
a thread-safe version of it called strerror_r(), but it is a bit clunky
to use and there are two different variants of it, with different
function prototypes.

On the other hand, strangely enough, perror() is thread-safe. Also, the
glibc "%m" printf extension which prints a textual value of errno is
also thread-safe, because, despite man printf(3) not saying so,
strerror_r() is being used in printf's implementation and not
strerror().

The strategy of this change is to use plain perror() where possible, aka
when we are directly printing the error returned by a system call. When
we are printing an error generated by an isochron function, that is
always negative, so we create a new pr_err() function which saves the
negation of that error code into the errno variable and uses the glibc
printf %m extension to print it.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>