isochron: send: use pthreads to create a dedicated sender thread
There are two distinct reasons why it is desirable to introduce POSIX
threads into isochron.

First, it has happened to me that PTP sync went bad in the middle of a
test, then recovered, and isochron did not catch this because it only
checks for sync status at the beginning and the end of a test (sending
packets during the test is too high-priority to do anything else). As
such, the data set was bogus, with things such as negative path delay.
It would be nice if we could monitor the sync status during the actual
test too, without blocking sending or incurring any additional delays,
but this is not possible with a single-threaded design.

Second, it would be nice if we could offload some of the work done in
the high-priority sending thread, for example polling for TX timestamps,
to a separate, lower-priority thread.

So this patch moves isochron towards that design, by trivially creating
a pthread for the sender which calls the existing run_nanosleep() loop
function. We no longer need to call sched_setattr(), since we can use
the pthread attribute functionality to the same effect.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>