isochron: send: monitor sync status during the test
To ensure that the logged timestamps are valid, we need to monitor the
ptpmon/sysmon offsets at all times, not just at the beginning and the
end of a test.

Now that we've delegated the task of sending packets to a dedicated
pthread which is currently trivially joined immediately after creation,
we can actually do some useful work in the main thread.

The design is as follows: the sender thread sets a bool variable to
notify when it has finished its work (either on errors or on success).
The main thread checks the sync status once per second until one of the
following occurs: either the sync offset exceeds the expected threshold
(case in which it requests the sender to stop), or the sender thread has
finished (case in which the data was valid).

If the data is deemed to be invalid, isochron restarts and attempts to
send packets again, from the beginning. Timestamps are only logged to
the file when the sync is okay.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>