isochron: send: collect tx timestamps in a separate thread
In the current isochron sender design, there is a fundamental problem in
that one cycle might not last as much as the next. Furthermore, we might
have an unbounded loop within a sender cycle.

It all has to do with the way in which we collect TX timestamps. We
expect to receive 3 TX timestamps per packet, and we need 6 syscalls for
them (including the poll). In principle, we don't need the timestamps
right away (and they aren't available right away, either), we just need
them at the end of the test. Right now just send as fast as we can and
go back to sleep as fast as we can.

But due to socket rmem constraints, we can't actually defer the entire
TX timestamp processing until the end of the test, because the kernel
won't buffer that much data for us.

So what we do is we poll the error queue after sending each packet, to
see whether there is any timestamp to consume right away. That's the
first problem right there, there is the overhead of one syscall even if
there is no timestamp to process. The idea is to not leave all
timestamps until the end. The second problem is that there isn't any
bound to how much work we do in the sender loop. If the kernel decides
to troll us and not give us any TX timestamp for the first 1000 packets,
then give us all timestamps after packet 1001, then we'll end up making
6000 syscalls in packet 1001's cycle, and the deadline for packet 1002
will most likely be missed. That's a design problem.

To address that, we can create another pthread whose only task is to
process TX timestamps. This makes the high-priority sender task bounded
and without any unrelated work being done in its context.

Note that we might now run into the opposite problem: if the sender
thread runs in a cycle that is too tight on a single CPU system, the TX
timestamp thread might starve and the socket error queue might get full
again. I don't know what to do about that, but on the other hand that
tight schedule might not have been possible in the first place, were it
for the TX timestamp work being done in the context of the sender thread.

The design of the synchronization between threads is this: the TX
timestamp thread is scheduled first, and it should go to sleep waiting
for the first TX timestamp. Then the sender thread is scheduled and it
starts sending packets. Since the hardware might drop packets for
whatever reason, we still need a timeout when polling for timestamps,
and now, the timeout is calculated as follows: we give the thread a base
timeout value of 1 second. But there is a race condition, since the
sender thread is created potentially after the tx timestamp thread has
started polling, and it might actually time out before the first packet
is sent. This is especially so, since the sender sleeps for at least one
second before proceeding to send the first packet. So we need to account
for this by increasing the TX timestamp timeout with one second, to two.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>