isochron: send: allow the user to specify the CPU affinity of the sender thread
I have noticed an interesting effect on NXP S32G274, which is that when
the main thread prints ptpmon/sysmon offsets to the console during a
test, this interferes quite badly with the wakeup times of the sender
thread, presumably due to the time spent with irqsoff in the kernel
linflexuart serial driver.

This patch allows the user to specify a CPU affinity for that task and
that task only. The use case is to use taskset isochron towards a
certain group of CPUs for the main thread and timestamping thread, and
to use --cpu-mask towards a separate, isolated CPU for the sender
thread.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>