isochron: log: only iterate over log from start to stop
For the simple use case where we want to print a simple packet from the
log, the overhead is currently quite high, because we iterate through
the entire data.

But due to the fact that we've changed the way in which packets are
logged, we can determine the position at which a packet could be found
just by knowing its seqid. So we can just iterate from the start seqid
to the stop seqid, instead of through the whole log.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>