isochron: use the "isochron.dat" file by default for the report
It may happen that the user forgets to specify the --output-file to
isochron-send, and starts a test that runs for several hours, only to
find out that nothing was logged.

Since isochron is pretty much useless without logs, make the program
always save its output when timestamping is requested, and the default
file name will be called "isochron.dat".

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>