isochron: log: escape the percent character (%%)
Escaping the percent sign in the isochron report printf format was
proclaimed too soon as useless. Since we can print free-form text per
packet, we can in fact print executable code in a language like Python,
to do stuff such as evaluate the delta between two timestamps. And since
such languages have printf-like format specifiers too, we should be able
to output text that contains the '%' character, so we need to escape it.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>