isochron: allow the interface to not be the same as the ptp4l interface
As explained here:
https://github.com/vladimiroltean/tsn-scripts/issues/8#issuecomment-999432505

there exist network interfaces which have VLAN filtering forced on
(cannot be disabled). To work on these, the user must create a VLAN
interface on top of the real interface, and use isochron on that.
In this case, isochron will see VLAN-untagged traffic (basically don't
specify --vid on the sender) and the VLAN will be inserted and stripped
by the kernel.

The problem is that currently, isochron expects that ptp4l runs on the
same interface as isochron, otherwise port state monitoring won't work.
So this practically means that ptp4l should be run in the same VLAN as
isochron, which is undesirable.

To lift that limitation, we should make the ptpmon port state
comparisons more flexible. Practically speaking, we should attempt to
resolve the real device behind a potential VLAN device, and allow
isochron and ptp4l to be on separate VLAN devices as long as the real
device is the same.

We do the communication with the kernel using the rtnetlink socket, and
we need to link with libmnl for that.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>