isochron: send: make --sync-threshold mandatory
It seems to be confusing to users that --sync-threshold defaults to 0 ns.
Especially since we need to monitor 4 servoes (local&remote ptpmon,
local&remote sysmon), it also seems to be pretty unlikely for all of
them to converge to 0 at the same time, and stay 0 for the entire
duration of the test.

So simply force the user to provide the required sync threshold as
argument.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>