8021cb demo: update to support external port links to swp2/swp3
Current CB demo using the internal port swp4 and eno2 as source. And
dedicate to assign the IP and MAC to eno2.
This patch still support the eno2 and swp4 port as data source. But also
add support external device as data source links to the swp2 or swp3
port.

Example extenting to 4 boards with parameter 't':

	./tsn-scripts/8021cb_ring.sh -i 1 -t 4

This patch makes that:
Set the hardware links as swp0/swp1 in the ring links. Originally the
swp4 port links to internal port eno2 and assigned an IP: 172.15.0.X.
But the swp2/swp3 ports are still free.
Base on the case, this patch makes swp2/swp3 as external ports which can
be links to devices. And assigned vid and ip mac_address for them.

The IP arrange would be like this(4 boards as example, the x in
172.15.0.x):
	swp4	swp2	swp3
board1:	1	5	9
board2:	2	6	10
board3:	3	7	11
board4:	4	8	12

Assigned VID are:
	swp4	swp2	swp3
board1:	101	105	109
board2:	102	106	110
board3:	103	107	111
board4:	104	108	112

The mac address assigned(the xx in the 00:04:9f:63:35:xx):
	swp4	swp2	swp3
board1:	17	21	25
board2:	18	22	26
board3:	19	23	27
board4:	20	24	28

Take device links to board1 swp2 port as example:
mac address set to: 00:04:9f:63:35:21
vlan id: 105
ip address: 172.15.0.5

Start the configure as example:
./8021cb.sh -t 4 -i 1

Signed-off-by: Po Liu <po.liu@nxp.com>