isochron: send: figure out the session start time from the main thread
We'd like for the TX timestamp thread to synchronize with the sender
thread (and wait until actual transmission is expected to start).

For that, we need to follow the same logic for determining the initial
wakeup time, both here, and there.

But that logic depends on a variable which is determined from the sender
thread. That variable must be known by both threads, so replace the
"now" from run_nanosleep with a prog->session_start which is determined
prior to calling prog_start_threads().

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>