isochron: send: don't let the base time drift away on session retries
When the sync status goes bad in the middle of a test, we kill all
threads, end the session and try again with a new one.

All is well, save for the fact that multiple calls to run_nanosleep()
are not idempotent, and they modify prog->base_time by adjusting it with
the prog->shift_time. So the base time would be shifted twice on the
second try, etc.

Fix this by having a local base_time variable in run_nanosleep(), and
the global prog->base_time will always contain the value provided by the
user.

Fixes: 2b79851489e4 ("isochron: send: monitor sync status during the test")
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>