ptpmon: print name of tried ports in case ptpmon_query_port_state_by_name finds nothing
When we have multiple ptp4l instances running, it may happen that the
user forgets that he/she needs to set up unique --uds_address values
for them, and match up those values with isochron's --unix-domain-socket.

What will happen is we'll get a bunch of these uninformative messages:

 ptpmon failed to query port state: No such device

To be helpful, let's allocate some memory to save the name of the ptp4l
ports that we've skipped, and print them to stderr in case we found no
match. That should make the user realize that isochron is talking to the
wrong ptp4l.

 Interface eno0 not found amount 1 ports reported by ptp4l: swp0
 ptpmon failed to query port state: No such device

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>