isochron: rcv: forward empty log to sender when timed out without any packet
The isochron receiver puts the sender's request for the packet log on
hold when not all packets have yet been received.

It also sets up an expiry timer to ensure that the sender won't wait
forever. This timer expires 5 seconds after the test begins, or after
the last packet is received.

If the sender sends packets for more than 5 seconds but none are
received by the receiver, the timer will have already expired by the
time the sender asks for the log. In that case, the receiver will put
the sender on hold, but it will have forgotten that there is no logic to
ever give some info back to the sender.

Fix this by detecting whether the timer has already expired when we
receive a request for the log, and if it has, don't put the sender on
hold, just give it an empty log, which makes it stop waiting
indefinitely for us.

Fixes: 3f84c085ede5 ("isochron: wait until all packets are received before forwarding the log")
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>