isochron: send: add an option to omit remote sync status
For kselftests, the sender and receiver run on the same system and share
the same PHC. There is a switch in the middle. PTP runs between the
sender port and the switch port attached to it. The switch is MASTER,
sender port is SLAVE.

This configuration is sufficient, since it also guarantees that the
receiver port is synchronized to the switch (therefore to the network
time). The receiver port is always synchronized to the sender port
(they share the same PHC). So it makes no sense to have a ptp4l instance
between the second switch port and the receiver port. Yet isochron
requires it, otherwise it complains that it cannot find a ptp4l port of
the same name as the receiver interface.

Suppress this requirement by adding a new option basically saying
"I know what I'm doing, it is safe to ignore the remote sync status".

By making this a separate option from --omit-sync, we still instantiate
a ptpmon and a sysmon, which is actually essential to running a test
which has network synchronization guarantees.

In fact, I considered whether to save the "omit_remote_sync" setting to
the isochron log or not, and decided not to. The reason why "omit_sync"
is saved is to know whether to bother making sense of the timestamp
diffs or not. But even with "omit_remote_sync" set, the expectation is
that the timestamps still fully make sense, it's just that running ptp4l
on the receiver interface isn't *necessary* for that.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>