isochron: be VRF-aware by using SO_BINDTODEVICE
David Ahern explains here:
https://static.sched.com/hosted_files/ossna2017/fe/vrf-tutorial-oss.pdf

that it is recommended that applications support IP addresses in the
form of:

 <ip-addr>%<if-name>

and use SO_BINDTODEVICE towards <if-name>

When <if-name> is a VRF, this automatically uses the FIB associated with
that l3mdev, and consequently only that VRF's routes.

As opposed to using "ip vrf exec vrf0 isochron send ..." where _all_ IP
sockets use the VRF's routing table, here we can selectively bind the
stats socket to one VRF, and the data socket to another.

In particular, this is relevant in the kselftests environment where the
isochron receiver and sender run on the same station (so the connection
to the stats socket can be established via 127.0.0.1), whereas the data
socket must go over the actual wire (since the wire is what we want to
test). This is impossible with "ip vrf exec".

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>