isochron: report: fix printing strings that aren't NULL-terminated
isochron_printf_one_packet() contains a bug and actually does not ensure
that the printf string isn't NULL terminated. It just ensures that it
is, if we aren't printing anything.

But when we are, the stuff we print advances buf_ptr, but no '\0' is
added at *buf_ptr, because more stuff is supposed to come. So it is only
by chance (zeroes on stack) that we print valid lines, and no junk that
comes afterwards.

Fix the bug by ensuring that all code paths move buf_ptr forward, and
terminate the string with a NULL at the end of replacing variables with
their values, not at the beginning.

Fixes: e49a4fd7b14c ("isochron: report: add a printf format specifier")
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>