isochron: report: print packets that are missing on RX too
For PSFP selftests where we make a switch deliberately drop out of band
packets, it is useful to debug the dropped packets and see their TX
timestamps, for example.

But we can't, because 'isochron report' skips past missing seqids in the
receiver log.

Fix that by reporting those RX timestamps as zero.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>