isochron: add support for orchestration
Orchestration refers to the ability of coordinating multiple senders
from a central node, which ensures that they are all synchronized within
the required threshold to the network.

To this end, the goal is to make a sender respond to management messages
in the same way as a receiver currently does (GET packet log, GET ptpmon
offset, GET sysmon offset), but extend the management TLVs such that a
sender is fully controllable remotely. The result is a sub-program that
starts with no role in the network, but only listens for remote
management commands. This is called a daemon.

The orchestrator parses an init file which contains the descriptions of
all senders, deduces their receivers, then connects to all stations,
monitors their sync status and decides when to instruct them to start
the test.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>