isochron: avoid use of strncpy
strncpy() simply bails out when copying a source string whose size
exceeds the destination string size, potentially leaving the destination
string unterminated.

There are 2 cases in the isochron code base: one is when we make use of
an if_name, which is limited to IFNAMSIZ. This comes from the argparser,
which has validated its length already.

The other is the UDS file path, for the ptpmon. That has been validated
by the argparser to not be longer than UNIX_PATH_MAX too. But since I'd
still like to keep the ptpmon code as fairly independent, it's still
good to validate once, explicitly, during ptpmon_create, and use the
plain strcpy afterwards.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>