isochron: log: use memcpy and memcmp when reading/writing the file magic
The isochron dat file begins with 8 bytes having the value of ISOCHRON.
This is a string which doesn't require a NULL terminator, since it is
fixed in size.

Currently there is a semi-bug where we require the first byte after
ISOCHRON to be zero, to serve as a NULL terminator for ISOCHRON, so that
isochron_log_load() can work. This is true right now, as the 4 bytes
after the magic are the file version in big endian, and that is
currently 4 (00 00 00 04).

In any case, use the bounded memcpy and memcmp to limit the reads and
writes of the file magic to precisely 8 bytes, what we need.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>