isochron: implement a central signal handler
An isochron daemon might get stuck in isochron_send_init_ptpmon() due to
the fact that ptp4l wasn't started, and killing it with SIGTERM won't
help.

This is because, although the prog_query_utc_offset() loop in there has
a "signal_received" check, that variable is actually replicated per
translation module. And we check the signal_received from send.o,
whereas the signal was received by daemon.o.

Fix this by implementing a central signal handler with a central
signal_received variable.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>