orchestrate: only instantiate sysmon and ptpmon if --omit-sync is false
Even if we later call isochron_update_sync_monitor_enabled(false), the
daemon has to call isochron_send_init_ptpmon(), and this queries the UTC
offset as a measure of figuring out whether ptp4l has been started.

But of course, the whole point of --omit-sync is to not require ptp4l,
so simply don't instantiate the daemon's ptpmon and sysmon if we're not
going to use them.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>