monitor port admin and link states
We are vulnerable to the link going up or down during a test and don't
do anything about it.

When we use ptpmon, we implicitly rely on ptp4l's PS_FAULTY state which
does track rtnl events. But when we don't, we have to check IFF_RUNNING
on the interface ourselves.

Also, the port may have not been brought up, and some drivers fail some
operations (like SIOCSHWTSTAMP) when they're down. It is confusing to
look at such errors when the cause could have been prevented, so just
check the admin state of the interface immediately after opening the
rtnetlink socket.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>