fix IPv4 management socket getting broken with CONFIG_IPV6=n
IPv4-mapped IPv6 adddresses have a key problem: they don't work when the
kernel is compiled without IPv6 support. So we still have to explicitly
fall back to PF_INET in server listenfd code.

This would be fine, but later when we create a stats_fd based on the
stats_listenfd and we print the client address using inet_ntop(), we
need to remember whether the listenfd socket was IPv4 or IPv6.

This escalates rather quickly, because we need to create a structure
that contains the listenfd and the family for the listenfd.
The result is the introduction of a socket structure that hides away the
IPv4 vs IPv6 details. This change uses the new socket structure for
management traffic.

Fixes: 0a8b5e1eac6e ("daemon: allow management connections over IPv6")
Fixes: 563b4d118668 ("rcv: properly support IPv6")
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>